MSGBXEN ed.

+++ source/repos/Laagspanningsnet/Laagspanningsnet/AansluitpuntVerwijderen.cs
@@ -3,6 +3,8 @@
  * Aanpassingen :
  *  - 20180317 :
  *      - .ico toegevoegd
+ *  - 20180508 :
+ *      - Messagebox ter bevestiging toegevoegd
  */
 using System;
 using System.ComponentModel;
@@ -35,12 +37,19 @@ namespace Laagspanningsnet
         // Er is op de OK knop geklikt.
         private void BtnOK_Click(object sender, EventArgs e)
         {
-            // Wis de aansluitpunt uit de database
-            _database.DeleteAansluitpunt(cmbAansluitpunt.Text);
+            DialogResult result = MessageBox.Show("Aansluitpunt " + 
+                                                  cmbAansluitpunt.Text +
+                                                  " volledig verwijderen?\n\nOok alle aansluitingen van "+
+                                                  cmbAansluitpunt.Text + " zullen verwijderd worden.", "Verwijderen?", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
+            {
+                // Wis de aansluitpunt uit de database
+                _database.DeleteAansluitpunt(cmbAansluitpunt.Text);
 
-            // sluit het venster
-            DialogResult = DialogResult.OK;
-            Close();
+                // sluit het venster
+                DialogResult = DialogResult.OK;
+                Close();
+            }
         }
 
         // Er is op de anuleer knop geklikt.
